pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "birdler"
version = "0.1.0"
description = "Birdler: command-line voice cloning tool using ChatterboxTTS"
authors = ["Rob Banagale <rob@banagale.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
chatterbox-tts = "*"
yt-dlp = {version = "*", optional = true}
youtube-dl = {version = "*", optional = true}

[tool.poetry.extras]
youtube = ["yt-dlp", "youtube-dl"]

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"
pytest = "^8.3.5"

[tool.pytest.ini_options]
addopts = "--import-mode=importlib"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = [
    ".venv",
    ".git",
    "__pycache__",
    ".mypy_cache",
    ".pytest_cache",
    "node_modules",
    "build", "dist",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint]
select = ["E", "F", "I"]
fixable = ["ALL"]
